requirements.txt
black==25.12.0
click==8.3.1
exceptiongroup==1.3.1
iniconfig==2.3.0
isort==7.0.0
librt==0.7.7
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
Pygments==2.19.2
pyright==1.1.408
pytest==9.0.2
pytokens==0.3.0
ruff==0.14.11
tomli==2.4.0
typing_extensions==4.15.0
Flask==3.0.3
flask-cors==4.0.1
Flask-Mail==0.10.0
flasgger==0.9.7.1
gunicorn==22.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
